+++ penna-yaml-config/src/main/java/module-info.java
@@ -1,9 +1,9 @@
+import penna.api.config.Provider;
 import penna.config.yaml.YamlConfigProvider;
 
 module penna.config.yaml {
     requires org.slf4j;
     requires transitive penna.api;
-    requires penna.core;
 
     // (Optional) support for Jackson
     requires static com.fasterxml.jackson.databind;
@@ -15,5 +15,5 @@ module penna.config.yaml {
     // (Optional) support for Snakeyaml Engine
     requires static org.snakeyaml.engine.v2;
 
-    provides penna.api.configv2.Provider with YamlConfigProvider;
+    provides Provider with YamlConfigProvider;
 }