+++ penna-core/src/test/java/penna/core/internals/TestContextPoolManager.java
@@ -2,7 +2,6 @@ package penna.core.internals;
 
 import penna.core.logger.guard.LevelGuard;
 import penna.core.sink.Sink;
-import penna.core.sink.TestSink;
 
 import java.util.function.Supplier;