+++ sample/src/main/java/org/example/Main.java
@@ -7,6 +7,8 @@ import java.time.Duration;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 
+
+
 public class Main {
     private static final Logger logger = LoggerFactory.getLogger(Main.class);
 
@@ -15,17 +17,17 @@ public class Main {
         var counter = new AtomicInteger();
 
         while (keepRunning.get()) {
-            Thread.ofVirtual().name("Runner").start(() -> {
-                logger.debug("Still running running");
-                logger.atTrace()
-                        .addArgument(counter::get)
-                        .log("Counter is at {}");
-                var next = counter.incrementAndGet();
-                if (next >= 256) {
-                    keepRunning.set(false);
-                }
-            });
-            Thread.sleep(Duration.ofMillis(250));
+            Thread.ofVirtual().name("Trace logger").start(() -> logger.atTrace()
+                    .addArgument(counter::get)
+                    .log("Counter is at {}"));
+            Thread.ofVirtual().name("Debug logger").start(() -> logger.debug("Still running running"));
+            Thread.ofVirtual().name("Info logger").start(() -> logger.info("Still running running"));
+            Thread.sleep(Duration.ofMillis(100));
+            var next = counter.incrementAndGet();
+            if (next >= 256) {
+                keepRunning.set(false);
+
+            }
         }
     }
 }