+++ CHANGELOG.md
@@ -3,10 +3,9 @@
 All notable changes to Penna will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) with an addendum:
-the PATCH component is omitted when its value is `0`.
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 0.8.0
 
 ### `penna-api`